+++ traktor/client/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Client"]
+
+from traktor.client.client import Client