+++ traktor/client/models.py
@@ -1,0 +1,20 @@
+__all__ = [
+    "RGB",
+    "ProjectCreateRequest",
+    "ProjectUpdateRequest",
+    "TaskCreateRequest",
+    "TaskUpdateRequest",
+    "TagCreateRequest",
+    "TagUpdateRequest",
+]
+
+
+from traktor.models import (
+    RGB,
+    ProjectCreateRequest,
+    ProjectUpdateRequest,
+    TaskCreateRequest,
+    TaskUpdateRequest,
+    TagCreateRequest,
+    TagUpdateRequest,
+)