+++ traktor/models/tag.py
@@ -1,6 +1,10 @@
-from sqlalchemy import orm
+from typing import Optional
 
-from traktor.models.model import Colored
+import slugify
+from sqlalchemy import orm
+from pydantic import BaseModel, validator
+
+from traktor.models.model import RGB, Colored
 from traktor.models.entry_tag import entry_tag_table
 
 
@@ -20,3 +24,31 @@ class Tag(Colored):
         return f"Tag(name={self.name}, color={self.color})"
 
     __repr__ = __str__
+
+
+class TagCreateRequest(BaseModel):
+    name: str
+    color: Optional[str] = None
+
+    @validator("color")
+    def validate_color(cls, value):
+        if value is None:
+            return RGB().hex
+        c = RGB(value)
+        return c.hex
+
+    @property
+    def slug(self):
+        return slugify.slugify(self.name)
+
+
+class TagUpdateRequest(BaseModel):
+    name: Optional[str] = None
+    color: Optional[str] = None
+
+    @validator("color")
+    def validate_color(cls, value):
+        if value is None:
+            return None
+        c = RGB(value)
+        return c.hex